Split in two backend URLs. Because:
Frontend needs to talk with the Backend, so it needs the docker network IP
(which it can get by using the container name "pucpos-b").

Once the frontend sends html+css+js to the user, the user needs to talk with the Backend.
So the user needs an IP that works from his point of view (he doesn't know "pucpos-b")